Saving items in Storage

+++ Assets/Scripts/TasksSystem/Requirements/ItemsInStorageTaskRequirement.cs
@@ -28,13 +28,13 @@ namespace ColonizationMobileGame.TasksSystem.Requirements
 
         private Progress GetProgress()
         {
-            IEnumerable<KeyValuePair<ItemType, uint>> filteredItemsInStorage = data.LevelData.ItemsInStorage
+            IEnumerable<KeyValuePair<ItemType, int>> filteredItemsInStorage = data.LevelData.ItemsInStorage
                 .Where(kvp => itemsRequired
                     .Select(r => r.Type)
                     .Contains(kvp.Key));
             int current = filteredItemsInStorage
-                .Where(kvp => (int) kvp.Value >= itemsRequired.SingleOrDefault(r => r.Type == kvp.Key)?.Required)
-                .Sum(kvp => (int) kvp.Value);
+                .Where(kvp => kvp.Value >= itemsRequired.SingleOrDefault(r => r.Type == kvp.Key)?.Required)
+                .Sum(kvp => kvp.Value);
             int required = itemsRequired.Sum(r => r.Required);
 
             return new Progress(current, required);
